Broken_Access_Control/IDOR/requirements.txt
Flask==2.0.1
mysql-connector-python

